feat: adicionando testes funcionais de tickets e adicionando estrutura de pastas
