+++ bugs.md
@@ -7,3 +7,18 @@ post não retorna reponsebody
 ct14 -> status code do swagger está errado ou da api
 ct17 -> permite a duplicação de nomes
 ct18 -> perimte alterar deixando campos vazios
+
+-----
+ct03 -> mensagem de retorno quando tentamos cadastrar assento 100 é diferente do swagger
+ct04 -> permite comprar mais de um ingresso por acento
+ct05 -> permite criar sem id do usuario
+ct06 -> permite criar showtime fora do formato, permite criar com id inválido
+ct07 -> permite criar com id inválido'
+ct09 -> requisições com id quebram a api
+ct00 -> permite criar com campos em branco
+permite criar assento com campo não reconhecido
+não tem como requisitar id de usuário
+
+
+----
+swagger incompleto, para request body e não possui responsebody para os status code
