+++ functionalTests/support/base/baseFaker.js
@@ -9,5 +9,19 @@ export class BaseFaker {
      showtimes: ["2024-08-26T13:38:37.241Z"] 
     };
   }
+  randomTicket(){
+    return{
+      price: `${Faker.number.int({ min: 0, max: 99 })}`,
+      seatNumber: `${Faker.number.int({ min: 0, max: 60 })}`,
+      showtime: "2024-08-26T16:26:00.077Z",
+    }
+  }
 }
 
+/* {
+  "movieId": "string",
+  "userId": "string",
+  "seatNumber": 0,
+  "price": 0,
+  "showtime": "2024-08-26T16:26:00.077Z"
+} */
