+++ functionalTests/tests/movies/create/[CT01]-createValidMovie.js
@@ -1,0 +1,32 @@
+import {
+  BaseRest,
+  BaseChecks,
+  ENDPOINTS,
+  testConfig,
+  BaseFaker,
+} from "../../../support/base/baseTest.js";
+
+export const options = testConfig.options.unitThresholds;
+const base_uri = testConfig.environment.hml.url;
+const baseRest = new BaseRest(base_uri);
+const baseChecks = new BaseChecks();
+const baseFaker = new BaseFaker();
+
+export default () => {
+  const movie = baseFaker.randomMovie();
+  const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
+  baseChecks.checkStatusCode(urlRes, 201);
+};
+
+export function teardown() {
+  // Resgatando Id dos filmes
+  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  const movies = JSON.parse(getMovie.body);
+  const moviesIds = movies.map((movie) => movie._id);
+
+  //Deletando filmes
+  moviesIds.forEach((movieId) => {
+    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
+    baseChecks.checkStatusCode(res, 200);
+  });
+}
