+++ functionalTests/tests/movies/create/[CT04]-createMovieInvalidData.js
@@ -1,0 +1,39 @@
+import {
+  BaseRest,
+  BaseChecks,
+  ENDPOINTS,
+  testConfig,
+  BaseFaker,
+} from "../../../support/base/baseTest.js";
+
+export const options = testConfig.options.unitThresholds;
+const base_uri = testConfig.environment.hml.url;
+const baseRest = new BaseRest(base_uri);
+const baseChecks = new BaseChecks();
+const baseFaker = new BaseFaker();
+
+const movie = {
+  title: "123",
+  description: "@$%",
+  launchdate: "não ter formato hora",
+  showtimes: ["não ter formato hora"],
+};
+
+export default () => {
+  const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
+  baseChecks.checkStatusCode(urlRes, 400);
+  console.log(urlRes.body);
+};
+
+export function teardown() {
+  // Resgatando Id dos filmes
+  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  const movies = JSON.parse(getMovie.body);
+  const moviesIds = movies.map((movie) => movie._id);
+
+  //Deletando filmes
+  moviesIds.forEach((movieId) => {
+    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
+    baseChecks.checkStatusCode(res, 200);
+  });
+}
