+++ functionalTests/tests/movies/read/[CT06]-listMovie.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -19,15 +19,8 @@ export function setup() {
 }
 
 export default () => {
-  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
-  const movies = JSON.parse(getMovie.body);
-  const moviesIds = movies.map((movie) => movie._id);
-
-  moviesIds.forEach((movieId) => {
-    const res = baseRest.getId(ENDPOINTS.MOVIES_ENDPOINT, movieId);
-    baseChecks.checkStatusCode(res, 200);
-  });
-
+  const urlRes = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  baseChecks.checkStatusCode(urlRes, 200);
 };
 
 export function teardown() {
