+++ functionalTests/tests/movies/read/[CT07]-searchMovieValidId.js
@@ -1,0 +1,43 @@
+import {
+  BaseRest,
+  BaseChecks,
+  ENDPOINTS,
+  testConfig,
+  BaseFaker,
+} from "../../../support/base/baseTest.js";
+
+export const options = testConfig.options.unitThresholds;
+const base_uri = testConfig.environment.hml.url;
+const baseRest = new BaseRest(base_uri);
+const baseChecks = new BaseChecks();
+const baseFaker = new BaseFaker();
+
+export function setup() {
+  const movie = baseFaker.randomMovie();
+  const resMovie = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
+  baseChecks.checkStatusCode(resMovie, 201);
+}
+
+export default () => {
+  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  const movies = JSON.parse(getMovie.body);
+  const moviesIds = movies.map((movie) => movie._id);
+
+  moviesIds.forEach((movieId) => {
+    const res = baseRest.getId(ENDPOINTS.MOVIES_ENDPOINT, movieId);
+    baseChecks.checkStatusCode(res, 200);
+  });
+};
+
+export function teardown() {
+  // Resgatando Id dos filmes
+  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  const movies = JSON.parse(getMovie.body);
+  const moviesIds = movies.map((movie) => movie._id);
+
+  //Deletando filmes
+  moviesIds.forEach((movieId) => {
+    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
+    baseChecks.checkStatusCode(res, 200);
+  });
+}
