+++ functionalTests/tests/movies/read/[CT08]-searchMovieInvalidId.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -13,8 +13,6 @@ const baseChecks = new BaseChecks();
 const baseFaker = new BaseFaker();
 
 export default () => {
-  let newMovie = baseFaker.randomMovie();
-  const res = baseRest.put(ENDPOINTS.MOVIES_ENDPOINT, newMovie, "123abc");
-  baseChecks.checkStatusCode(res, 201);
-  console.log(res.body);
+  const resId = baseRest.getId(ENDPOINTS.MOVIES_ENDPOINT, "123abc");
+  baseChecks.checkStatusCode(resId, 400);
 };
