+++ functionalTests/tests/movies/update/[CT15]-editMovieValidId.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -12,8 +12,9 @@ const baseRest = new BaseRest(base_uri);
 const baseChecks = new BaseChecks();
 const baseFaker = new BaseFaker();
 
-
 export default () => {
-    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, "123abc");
-    baseChecks.checkStatusCode(res, 404)
+  let newMovie = baseFaker.randomMovie();
+  const res = baseRest.put(ENDPOINTS.MOVIES_ENDPOINT, newMovie, "123abc");
+  baseChecks.checkStatusCode(res, 201);
+  console.log(res.body);
 };
