+++ functionalTests/tests/movies/update/[CT18]-editMovieWithEmptyFields.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -14,11 +14,30 @@ const baseFaker = new BaseFaker();
 
 export function setup() {
   const movie = baseFaker.randomMovie();
-  const resMovie = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
-  baseChecks.checkStatusCode(resMovie, 201);
+  const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
+  baseChecks.checkStatusCode(urlRes, 201);
 }
 
+const newMovie = {
+  title: "Ephraim47",
+  description: " ",
+  launchdate: " ",
+  showtimes: [" "],
+};
+
 export default () => {
+  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
+  const movies = JSON.parse(getMovie.body);
+  const moviesIds = movies.map((movie) => movie._id);
+
+  moviesIds.forEach((movieId) => {
+    const res = baseRest.put(ENDPOINTS.MOVIES_ENDPOINT, newMovie, movieId);
+    baseChecks.checkStatusCode(res, 201);
+    console.log(res.body);
+  });
+};
+
+export function teardown() {
   // Resgatando Id dos filmes
   const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
   const movies = JSON.parse(getMovie.body);
@@ -29,4 +48,4 @@ export default () => {
     const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
     baseChecks.checkStatusCode(res, 200);
   });
-};
+}
