+++ functionalTests/tests/tickets/create/[CT03]-createTicketInvalidPrice.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -12,21 +12,29 @@ const baseRest = new BaseRest(base_uri);
 const baseChecks = new BaseChecks();
 const baseFaker = new BaseFaker();
 
-export default () => {
+export function setup() {
   const movie = baseFaker.randomMovie();
   const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
   baseChecks.checkStatusCode(urlRes, 201);
-};
+}
 
-export function teardown() {
-  // Resgatando Id dos filmes
+export default () => {
   const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
   const movies = JSON.parse(getMovie.body);
   const moviesIds = movies.map((movie) => movie._id);
 
-  //Deletando filmes
+  //cadastrando tickets
   moviesIds.forEach((movieId) => {
-    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
-    baseChecks.checkStatusCode(res, 200);
+    const ticket = {
+      movieId: `${movieId}`,
+      userId: "string",
+      seatNumber: 12,
+      price: 230,
+      showtime: "2024-08-26T16:26:00.077Z",
+    };
+
+    const res = baseRest.post(ENDPOINTS.TICKETS_ENDPOINT, ticket);
+    baseChecks.checkStatusCode(res, 400);
+    console.log(res.body);
   });
-}
+};
