+++ functionalTests/tests/tickets/create/[CT07]-createTicketInvalidMovie.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -12,29 +12,28 @@ const baseRest = new BaseRest(base_uri);
 const baseChecks = new BaseChecks();
 const baseFaker = new BaseFaker();
 
-const movie = {
-  title: "123",
-  description:
-    "@$%",
-  launchdate: "não ter formato hora",
-  showtimes: ["não ter formato hora"],
-};
+export function setup() {
+  const movie = baseFaker.randomMovie();
+  const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
+  baseChecks.checkStatusCode(urlRes, 201);
+}
 
 export default () => {
-  const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
-  baseChecks.checkStatusCode(urlRes, 400);
-  console.log(urlRes.body);
-};
-
-export function teardown() {
-  // Resgatando Id dos filmes
   const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
   const movies = JSON.parse(getMovie.body);
   const moviesIds = movies.map((movie) => movie._id);
 
-  //Deletando filmes
+  //cadastrando tickets
   moviesIds.forEach((movieId) => {
-    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
-    baseChecks.checkStatusCode(res, 200);
+    const ticket = {
+      movieId: `123abc`,
+      userId: "string",
+      seatNumber: 4,
+      price: 23,
+      showtime: "fora do formato",
+    };
+
+    const res = baseRest.post(ENDPOINTS.TICKETS_ENDPOINT, ticket);
+    baseChecks.checkStatusCode(res, 400);
   });
-}
+};
