+++ functionalTests/tests/tickets/delete/[CT13]-deleteTicketValidId.js
@@ -4,7 +4,7 @@ import {
   ENDPOINTS,
   testConfig,
   BaseFaker,
-} from "../../support/base/baseTest.js";
+} from "../../../support/base/baseTest.js";
 
 export const options = testConfig.options.unitThresholds;
 const base_uri = testConfig.environment.hml.url;
@@ -16,37 +16,31 @@ export function setup() {
   const movie = baseFaker.randomMovie();
   const urlRes = baseRest.post(ENDPOINTS.MOVIES_ENDPOINT, movie);
   baseChecks.checkStatusCode(urlRes, 201);
-}
 
-const newMovie = {
-  title: "Ephraim47",
-  description:
-    " ",
-  launchdate: " ",
-  showtimes: [" "],
-};
-
-export default () => {
   const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
   const movies = JSON.parse(getMovie.body);
   const moviesIds = movies.map((movie) => movie._id);
 
+  //cadastrando tickets
   moviesIds.forEach((movieId) => {
-    const res = baseRest.put(ENDPOINTS.MOVIES_ENDPOINT, newMovie, movieId);
+    const ticket = {
+      movieId: `${movieId}`,
+      userId: "string",
+      seatNumber: 4,
+      price: 23,
+      showtime: "2024-08-26T16:26:00.077Z",
+    };
+
+    const res = baseRest.post(ENDPOINTS.TICKETS_ENDPOINT, ticket);
     baseChecks.checkStatusCode(res, 201);
-    console.log(res.body);
   });
-};
 
-export function teardown() {
-  // Resgatando Id dos filmes
-  const getMovie = baseRest.get(ENDPOINTS.MOVIES_ENDPOINT);
-  const movies = JSON.parse(getMovie.body);
-  const moviesIds = movies.map((movie) => movie._id);
-
-  //Deletando filmes
-  moviesIds.forEach((movieId) => {
-    const res = baseRest.delete(ENDPOINTS.MOVIES_ENDPOINT, movieId);
-    baseChecks.checkStatusCode(res, 200);
-  });
+  return {
+    id: moviesIds,
+  };
 }
+
+export default (id) => {
+  const urlRes = baseRest.delete(ENDPOINTS.TICKETS_ENDPOINT, id);
+  baseChecks.checkStatusCode(urlRes, 200);
+};
